Lecture02/notebooks/4_full_project/files/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
pyyaml>=6.0
matplotlib>=3.5.0
pandas>=1.4.0
